@@ -2,10 +2,13 @@
   "name": "sd-modal",
   "version": "0.0.1",
   "description": "A lightweight package for using modals in a vue project.",
-  "main": "dist/sd-modal.common.js",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "unpkg": "dist/vuelib.min.js",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build --target lib ./src/index.ts",
+    "build-rollup": "rimraf dist && cross-env NODE_ENV=production rollup -c",
     "lint": "vue-cli-service lint",
     "release": "standard-version",
     "prepare": "husky install"
@@ -13,14 +16,17 @@
   "keywords": [],
   "author": "Kevin Izuchukwu",
   "license": "ISC",
-  "dependencies": {
-    "@vue/composition-api": "^1.0.0-rc.12",
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11"
-  },
   "devDependencies": {
+    "@babel/preset-env": "^7.14.5",
     "@commitlint/cli": "^12.1.4",
     "@commitlint/config-conventional": "^12.1.4",
+    "@rollup/plugin-alias": "^3.1.2",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^19.0.0",
+    "@rollup/plugin-image": "^2.0.6",
+    "@rollup/plugin-node-resolve": "^13.0.0",
+    "@rollup/plugin-replace": "^2.4.2",
+    "@rollup/plugin-url": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
     "@typescript-eslint/parser": "^4.18.0",
     "@vue/cli-plugin-babel": "~4.5.0",
@@ -29,11 +35,24 @@
     "@vue/cli-service": "~4.5.0",
     "@vue/eslint-config-typescript": "^7.0.0",
     "babel-eslint": "^10.1.0",
+    "core-js": "^3.15.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
     "husky": "^6.0.0",
+    "postcss-import": "^14.0.2",
+    "postcss-nested": "^5.0.5",
+    "postcss-simple-vars": "^6.0.3",
+    "postcss-url": "^10.1.3",
+    "rollup": "^2.52.1",
+    "rollup-plugin-css-only": "^3.1.0",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-postcss": "^4.0.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup-plugin-vue": "^6.0.0",
     "standard-version": "^9.3.0",
     "typescript": "~4.1.5",
+    "vue": "^3.1.1",
     "vue-template-compiler": "^2.6.11"
   },
   "browserslist": [
