@@ -1,6 +1,4 @@
-import Vue from 'vue'
+import { createApp  } from 'vue'
 import { HelloWorld } from './index'
 
-new Vue({
-	render: (h) => h(HelloWorld),
-}).$mount('#app')
+createApp(HelloWorld).mount('#app')
